+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # develop
   * Feature: Support outbound dial on Asterisk
+  * Bugfix: Asterisk hangup causes should map to correct Rayo End reason
 
 # v0.8.1 - 2012-01-09
   * Feature: Support DTMF Input components on Asterisk